Developed service to validate owner role for restaurant creation

+++ src/main/java/com/plazoleta/users/infrastructure/output/jpa/repository/IUserRepository.java
@@ -3,7 +3,7 @@ package com.plazoleta.users.infrastructure.output.jpa.repository;
 import com.plazoleta.users.infrastructure.output.jpa.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IUserRepository extends JpaRepository<UserEntity,Integer> {
+public interface IUserRepository extends JpaRepository<UserEntity,Long> {
     boolean existsByEmail(String email);
     boolean existsByDocumentNumber(String documentNumber);
 }
